requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-multipart==0.0.18

# AI and LLM dependencies
crewai==0.86.0
crewai-tools==0.17.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-anthropic==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0

# Agent tools and integrations
composio-crewai==0.7.0
requests==2.32.3

# Data and utilities
pyyaml==6.0.1
python-dotenv==1.0.1
aiofiles==24.1.0

# Vector storage for memory
pinecone-client==5.0.0
faiss-cpu==1.8.0
numpy==2.0.0

# GitHub integration
PyGithub==2.4.0
gitpython==3.1.40

# Development and testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.27.0

# Code quality
flake8==7.1.0
mypy==1.11.0
black==24.8.0
isort==5.13.0
bandit==1.8.0

# Documentation
mkdocs==1.6.0
mkdocs-material==9.5.0
mkdocstrings[python]==0.26.0
mkdocs-gen-files==0.5.0
mkdocs-literate-nav==0.6.0
